requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pillow>=10.0.0
openai>=1.3.0
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0
google-cloud-texttospeech>=2.16.0